+++ lib/trunk/src/pagespeed/image_compression/png_optimizer.cc
@@ -42,8 +42,8 @@ void ReadPngFromStream(png_structp read_ptr,
                        png_bytep data,
                        png_size_t length) {
   PngInput* input = reinterpret_cast<PngInput*>(read_ptr->io_ptr);
-  int copied = input->data_->copy(reinterpret_cast<char*>(data), length,
-                                  input->offset_);
+  size_t copied = input->data_->copy(reinterpret_cast<char*>(data), length,
+                                     input->offset_);
   input->offset_ += copied;
   if (copied < length) {
     png_error(read_ptr, "ReadPngFromStream: Unexpected EOF.");
