+++ lib/trunk/src/pagespeed/rules/minimize_request_size_test.cc
@@ -114,9 +114,9 @@ class MinimizeRequestSizeTest : public ::testing::Test {
 
     {
       // Check contents of results object
-      ASSERT_EQ(results.results_size(), expected.size());
+      ASSERT_EQ(static_cast<size_t>(results.results_size()), expected.size());
 
-      for (int idx = 0; idx < expected.size(); ++idx) {
+      for (size_t idx = 0; idx < expected.size(); ++idx) {
         const Result& result = results.results(idx);
         ASSERT_EQ(result.resource_urls_size(), 1);
         EXPECT_EQ(expected[idx], result.resource_urls(0));
